Add new options package
